runner/resultgen: More checking if subtests should exist
If a test with subtests just exits immediately, or the test binary
doesn't exist at all (as is sometimes the case with kms_chamelium),
the existence of subtests doesn't end up in the execution journal. As
was done for timeouts in a797cbf6918a ("runner/resultgen: Be more
robust with incomplete tests"), check if we were attempting to run a
subtest before attributing a 'notrun' result to an incorrect field.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Cc: Martin Peres <martin.peres@linux.intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>