runner: Add time_elapsed field to results
When starting a test run, drop a timestamp file. Do the same when
ending a run. Slap those timestamps directly into the time_elapsed
field in results.json.

Using timestamps instead of measuring actual elapsed time goes against
the naming of the field, but the name is chosen by piglit. Even though
piglit itself uses timestamps.

Corner cases:

On incomplete test runs, the end timestamp will be missing. The
time_elapsed field will only have the start timestamp. This matches
piglit behaviour exactly.

On incomplete but resumed test runs, the end timestamp will be the
time when the resume finishes. Piglit doesn't do this, and instead
leaves the end timestamp missing. Discussing which behaviour is better
is left as an exercise to the readers.

Signed-off-by: Petri Latvala <petri.latvala@intel.com>
Cc: Arkadiusz Hiler <arkadiusz.hiler@intel.com>
Cc: Tomi Sarvela <tomi.p.sarvela@intel.com>
Reviewed-by: Arkadiusz Hiler <arkadiusz.hiler@intel.com>