igt/pm_rpm: Ignore modesets for basic tests with no KMS
If KMS is not available, we cannot simply turn on an output and expect
that to wake the device up. As such we have to ignore that part of the
basic subtest and simply proclaim victory if the device is able to
sleep!

v2: Mention that we can test wakeup via nop execbuf in future.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>