gitlab-ci: Build docs by default
I just realized that when you push to a non-HEAD branch, then gitlab
doesn't bother to build the deploy stage. Which means if you push to a
branch, then that's not properly tested.

So build docs right away. We're not going to waste anything, since the
deploy:pages job will pick up right from where we left.

Reviewed-by: Eric Anholt <eric@anholt.net>
Reviewed-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>