tests: Make a fb-tiling subtest to check for invalid tiling changes.
In 2013 we prevented changing the tiling of an active framebuffer:

commit 80075d492f8773209e26d11d6bb13ba624ef95a4
Author: Daniel Vetter <daniel.vetter@ffwll.ch>
Date:   Wed Oct 9 21:23:52 2013 +0200

    drm/i915: prevent tiling changes on framebuffer backing storage

After this we no longer have any use for the bad-tiling subtest, and
only have to make sure changing tiling on a fb is not allowed.

Remove the original bad-tiling subtest, and add a i915_fb_tiling
subtest for this case.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
[mlankhorst: Fix build system changes]
Reviewed-by: Juha-Pekka Heikkila <juhapekka.heikkila@gmail.com>