tests/i915/gem_render_tiled_blits: Make sure igt_buf is initially zero'd
gen8_bind_buf looks at buf->aux.stride/offset, which is not set by the
test. When I added the bpp field, it was enough to make the test fail,
fix this by making sure that the buf is initially zero'd explicitly,
which fixes the test to run reliably 100% of the time.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Reviewed-by: Petri Latvala <petri.latvala@intel.com> #irc