gitlab-ci: Always set PKG_CONFIG_PATH
meson gets grumpy if that changes between invocations.

I'm not entirely clear on why this happens with the interim runner
from gnome folks, but it does. We do still run the exact same image,
with the exact same meson version. But someone the old runner didn't
complain, while the new one does.

Note that the jobs still fail because the interim runner is dead slow
(10x-100x slower to run the library tests) and times out after 30s on
tests that take 1s max.

Acked-by: Petri Latvala <petri.latvala@intel.com>
Signed-off-by: Daniel Vetter <daniel.vetter@intel.com>