lib/igt_fb: Add a check on stride alignment for pixman conversion
Pixman requires buffer strides the be aligned to 32-bit words in order
to create internal representations of buffers. If this condition is not
met, it will fail and IGT will not be able to report the error cause,
making it hard to debug the issue.

Add an explicit check in our code prior to calling pixman when
converting buffer so that the error can be understood if it occurs.

Signed-off-by: Paul Kocialkowski <paul.kocialkowski@bootlin.com>
Reviewed-by: Maxime Ripard <maxime.ripard@bootlin.com>