i915/gem_tiled_pread: Skip on unknown swizzling
If we do not know the underlying swizzle on the HW, we do not know the
full tiling pattern and cannot predict the expected results. This is
often because the swizzle varies between pages and is not as constant as
we naively expected.

v2: gem_get_tiling() does the physical==reported check, we just need to
add a require

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc:  Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@linux.intel.com>